KingdomsSimulator - Mem optimalizations and maintenance
 - still not verified it is running
 - still not all leaks resolved

+++ objc/KingdomsSimulator/KingdomsSimulator/CommandLine.h
@@ -10,11 +10,11 @@
 
 @interface CommandLine : NSObject
 {
-    NSInteger cyclesCount;
+    int cyclesCount;
     NSString* configPath;
 }
 
-@property(readonly) NSInteger cyclesCount;
+@property(readonly) int cyclesCount;
 @property(readonly) NSString* configPath;
 
 - (id) initWithArgc: (int) argc andArgv: (const char*[]) argv;
